Fixed old recompiler compiling in a kludgy way because it appears there's no STORE_IMM_ADDR_W.

+++ src/cpu/cpu.h
@@ -784,7 +784,7 @@ typedef struct {
     uint32_t smhr;
 } cyrix_t;
 
-extern uint16_t x87_op;
+extern uint32_t x87_op;
 
 extern uint32_t addr64;
 extern uint32_t addr64_2;
